+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Nothing yet
+- More test cases added.
 
 ### Changed
 
-- Nothing yet
+- Rust default compiler settings changed to include `-C opt-level=3`.
 
 ### Fixed